Client-side logic for writing to the database.
index.html: basic UI for the Blockly workspace.

index.js: Entry point for the client. Includes a change listener that
detects events created on the UI and logic to send those changes to the
server.

WorkspaceClient.js: Class for tracking events and handling event
resolution. This change adds support for tracking events created on the
client and writing them to the database.