Batch events for writing to the database. (#21)
Currently, each event is written to the database as an individual entry.
With this change, the client batches any available events and writes
them to the database as a single entry. By writing all events to the
database in a single entry this removes the need to use a transaction.

Other:
WorkspaceClient.js: this.inProgress was an array of events, now
this.inProgress is an array of LocalEntry objects. Tracking a LocalEntry
object facilitates merging client and server events. Currently, only one
LocalEntry event will exist at a time, however, once websockets are
implemented, inProgress will hold multiple LocalEntry objects.