Store client entryIds to validate database writes. (#32)
A second table in the database will be used to track information about
each client. Current use is for keeping track of the last entryId of
each client. The server will use this information to verify if an entry
it received is in the correct order before adding it to the database.