Fix procedure var names being uneditable from XML
Fixed not being able to edit procedure var names after the procedure
was loaded from XML. This was occuring because the variable would be
created with a different ID by the workspace, so the id passed to
the field would be incorrect. When the field tried to access the var
based on its name, it would get a null return value.