fix(dev-tools): Have runSerializationTestSuite accept Blockly
Have runSerializationTestSuite accept a second argument which is
the Blockly object to use for XML serialization/deserialization.

This works around an issue in blockly-samples that (following
the deletion of the alias for blockly in webpack.config.js)
causes tests using this function to fail due to having two copies
of jsdom loaded, where each will reject DOM objects created by
the other.  See
https://github.com/google/blockly-samples/pull/2229#issuecomment-1979123919
for more details.