fix: dev-tools only set monaco model and text when changing
Context:
- https://github.com/mit-cml/workspace-multiselect/pull/62#issuecomment-2196819235
- https://github.com/mit-cml/workspace-multiselect/pull/62#issuecomment-2212561339
- https://github.com/mit-cml/workspace-multiselect/pull/62#issuecomment-2214203536

In multi-select plugin, a significant amount of workspace
change events (for each block) can be triggered at the same
time, which causes the garbage collection mechanism to fail,
so that we eventually have JS heap overflow and the page
crashed in Chromium.

This commit introduces a workaround for this, so that we only
update the page when we have model or text updates in monaco.

Signed-off-by: Hollow Man <hollowman@opensuse.org>