fix(build): Use current version of GitHub actions to avoid warnings (#2461)
The old action versions previously specified attempted to use Node.js
v12, which is no longer supported, and a warning is issued when
they run on v16 instead.  Update build.yml so that we use versions
of the actions that use a version of node.js that is still supported.