fix!: Delete (old) keyboard-navigation plugin
This depends on things that no longer exist in
blockly@v12.0.0-beta.4, like BasicCursor and Cursor, and also
on other things that will cease to exist before 12.0.0, like
ASTNode.

We need to update the samples repository to be able to test
plugins with v12, and this plugin is soon to be replaced by
the new plugin being developed in
google/blockly-keyboard-experimentation, so the most expedient
course of action is just to delete this one to allow upgrading
this repo's blockly devDependency.

(We could alternatively leave this plugin as an empty shell
waiting to be filled with the contents of the
blockly-keyboard-experimentation repo, but it seems preferable
to not publish publish anything, thus leaving the latest version
of the @blockly/keyboard-navigation NPM package useful though
outdatedly dependent on blockly@11, rather than publish an empty
plugin package that is of no use at all.

BREAKING CHANGE:

Removes the keyboard-experimentation plugin, obviously.