fix!: Delete old keyboard-navigation plugin (#2508)
This depends on things that no longer exist in
blockly@v12.0.0-beta.4, like BasicCursor and Cursor, and also
on other things that will cease to exist before 12.0.0, like
ASTNode.

We need to update the samples repository to be able to test
plugins with v12, and this plugin is soon to be replaced by
the new plugin being developed in
google/blockly-keyboard-experimentation, so the most expedient
course of action is just to delete this one to allow upgrading
this repo's blockly devDependency.

BREAKING CHANGE:

Removes the keyboard-experimentation plugin, obviously.