feat(workspace-search): focus found block on close or restore focus
Uses Blockly 12 FocusManager API.

This matches browser in find in page behaviour where Escape closes find and
moves focus to the highlighted item.

When there's no search performed or nothing ever matched we restore focus to
the previous node.

Where there was a match but then the search changes so it no longer matched we
still focus that last match, because we've scrolled it into view which may have
moved the previous focus position out of view. In this scenario browsers seem
to focus the element but not mark it as focus:visible but we don't have a
similar distinction.